requirements.txt
numpy ==  1.23.1
tensorflow-cpu==2.9.1
opencv-python-headless==4.4.0.42
streamlit ==  1.11.0
Pillow== 9.2.0
scikit-learn==0.23.1
pandas == 1.4.3
pickleshare ==  0.7.5
keras == 2.9.0
requests ==  2.24.0
tensorflow-hub == 0.12.0
